Move data sources map endpoint to `/map/data-sources`
